requirements.txt
click==8.1.8
dlib==19.24.6
face-recognition==1.3.0
face_recognition_models==0.3.0
numpy==1.24.3
opencv-contrib-python==4.8.0
opencv-python-headless==4.11.0.86
pillow==10.0.0
matplotlib==3.7.2
tk==0.1.0
